requirements.txt
fastapi[standard]
llama-index-core
llama-index-llms-openai-like
fastembed
llama-index-embeddings-openai
python-dotenv
psycopg2-binary
sqlalchemy
pgvector
requests
yfinance
edgartools
hishel<0.1.0
sentence-transformers